refactoring newpost and put all related data and functions in it and removing from content
