+++ data/db.json
@@ -14,21 +14,9 @@
     },
     {
       "id": 3,
-      "title": "Saeed Njafi",
-      "datetime": "July 21, 2022 12:57:10 PM",
+      "title": "Saeed Njaf",
+      "datetime": "July 21, 2022 5:03:47 PM",
       "body": "this is also modified by me as i hate lerom text "
-    },
-    {
-      "id": 5,
-      "title": "test2",
-      "datetime": "July 21, 2022 12:55:53 PM",
-      "body": "my start post \nsecond \n"
-    },
-    {
-      "id": 6,
-      "title": "ad05",
-      "datetime": "July 21, 2022 4:39:36 PM",
-      "body": "sf;km"
     }
   ]
 }
