+++ src/Header.js
@@ -1,10 +1,9 @@
 import React from "react";
 import {FaLaptop ,FaTabletAlt, FaMobileAlt } from 'react-icons/fa';
-import DataContext from "./context/DataContext";
-import { useContext } from "react";
+import useWindowSize from "./hooks/useWindowSize";
 
 const Header = ({title}) => {
-  const {width} = useContext(DataContext);
+  const {width} =useWindowSize();
 
   return (
     <header className="Header">
