+++ src/NewPost.js
@@ -1,9 +1,32 @@
 import React from "react";
-import { useContext } from "react";
+import { useContext , useState} from "react";
 import DataContext from "./context/DataContext";
+import { useNavigate } from "react-router-dom";
+import { format } from "date-fns";
+import api from './api/posts'
 
 const NewPost = () => {
-  const {handleSubmit, postBody, setPostBody , postTitle , setPostTitle} = useContext(DataContext);
+  const [postTitle, setPostTitle] = useState("");
+  const [postBody, setPostBody] = useState("");
+  const {posts, setPosts} = useContext(DataContext);
+  const navigate = useNavigate();
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    const id = posts.length ? posts[posts.length - 1].id + 1 : 1;
+    const datetime = format(new Date(), "MMMM dd, yyyy pp");
+    const newPost = { id, title: postTitle, datetime, body: postBody };
+    try {
+      const response = await api.post("./posts", newPost);
+      const allPosts = [...posts, response.data];
+      setPosts(allPosts);
+      setPostTitle("");
+      setPostBody("");
+      navigate("/");
+    } catch (err) {
+      console.log(`Error: ${err.message}`);
+    }
+  };
   
   return (
     <main className="NewPost">
