+++ src/context/DataContext.js
@@ -2,7 +2,6 @@ import { useEffect, useState, createContext } from "react";
 import { Route, Routes, useNavigate } from "react-router-dom";
 import { format } from "date-fns";
 import api from "../api/posts";
-import useWindowSize from "../hooks/useWindowSize";
 import useAxiosFetch from "../hooks/useAxiosFetch";
 
 const DataContext = createContext({});
@@ -11,13 +10,10 @@ export const DataProvider = ({ children }) => {
   const [posts, setPosts] = useState([]);
   const [search, setSearch] = useState("");
   const [searchResults, setSearchResults] = useState([]);
-  const [postTitle, setPostTitle] = useState("");
-  const [postBody, setPostBody] = useState("");
+ 
   const [editTitle, setEditTitle] = useState("");
   const [editBody, setEditBody] = useState("");
-
   const navigate = useNavigate();
-  const { width } = useWindowSize();
   const { data, fetchError, isLoading } = useAxiosFetch(
     "http://localhost:3500/posts"
   );
@@ -54,22 +50,7 @@ export const DataProvider = ({ children }) => {
     setSearchResults(filteredResults.reverse());
   }, [posts, search]);
 
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    const id = posts.length ? posts[posts.length - 1].id + 1 : 1;
-    const datetime = format(new Date(), "MMMM dd, yyyy pp");
-    const newPost = { id, title: postTitle, datetime, body: postBody };
-    try {
-      const response = await api.post("./posts", newPost);
-      const allPosts = [...posts, response.data];
-      setPosts(allPosts);
-      setPostTitle("");
-      setPostBody("");
-      navigate("/");
-    } catch (err) {
-      console.log(`Error: ${err.message}`);
-    }
-  };
+  
 
   const handleEdit = async (id) => {
     const datetime = format(new Date(), "MMMM dd, yyyy pp");
@@ -97,15 +78,25 @@ export const DataProvider = ({ children }) => {
     }
   };
   return (
-    <DataContext.Provider value={{
-        width , search , setSearch, searchResults, fetchError,
-        isLoading, handleSubmit, postBody, setPostBody, postTitle,
-        setPostTitle, posts, handleEdit, editBody, setEditBody,
-        editTitle, setEditTitle
-    }}>
-        {children}
+    <DataContext.Provider
+      value={{
+        search,
+        setSearch,
+        searchResults,
+        fetchError,
+        isLoading,
+        posts,
+        handleEdit,
+        editBody,
+        setEditBody,
+        editTitle,
+        setEditTitle,
+        handleDelete,
+      }}
+    >
+      {children}
     </DataContext.Provider>
-);
+  );
 };
 
 export default DataContext;
